# quizano/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Quizano Simple App API
  version: 1.0.0
  description: Node.js API for a simplified Quizano app with student and admin flows.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users
      description: Admin only. Filter by role and search keyword.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [student, admin]
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/users/students:
    post:
      tags: [Users]
      summary: Create student user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
      responses:
        '201':
          description: Student created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/users/{userId}:
    put:
      tags: [Users]
      summary: Update user profile
      description: Admin only. Update fullName, email, password.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Users]
      summary: Delete user
      description: Admin only. Cannot delete admin user.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
                      userId:
                        type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/results/submit:
    post:
      tags: [Results]
      summary: Submit exam answers
      description: Student only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitResultRequest'
      responses:
        '201':
          description: Submission accepted and graded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      resultId:
                        type: string
                      score:
                        type: number
                      correctCount:
                        type: integer
                      totalQuestions:
                        type: integer
                      status:
                        type: string
                        example: completed
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/results/me:
    get:
      tags: [Results]
      summary: List current student's results
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: examId
          schema:
            type: string
      responses:
        '200':
          description: Results list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Result'
  /api/v1/results/me/latest:
    get:
      tags: [Results]
      summary: Latest result of current student
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Latest result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Result'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/results/{resultId}:
    get:
      tags: [Results]
      summary: Get result by id
      description: Student can read own result, admin can read any.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: resultId
          required: true
          schema:
            type: string
        - in: query
          name: includeDetails
          schema:
            type: boolean
      responses:
        '200':
          description: Result detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/Result'
                      - $ref: '#/components/schemas/ResultDetailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/admin/students/{studentId}/results/{resultId}:
    get:
      tags: [Admin]
      summary: Admin view detailed result of one student
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
        - in: path
          name: resultId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AdminDetailedResult'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Validation failed
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Invalid or expired access token
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Resource not found
    Conflict:
      description: Conflict state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Username already exists
  schemas:
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            details:
              nullable: true
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
...
    RegisterRequest:
      type: object
      required: [fullName, username, email, password]
      properties:
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 3
...
    CreateStudentRequest:
      allOf:
        - $ref: '#/components/schemas/RegisterRequest'
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
    Exam:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [free, scheduled]
        duration:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        startTime:
          type: string
          format: date-time
          nullable: true
        endTime:
          type: string
          format: date-time
          nullable: true
...
    Option:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
...
    AnswerInput:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: string
        selectedOptionId:
          type: string
          nullable: true
    SubmitResultRequest:
      type: object
      required: [examId, answers]
      properties:
        examId:
          type: string
        startTime:
          type: string
          format: date-time
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerInput'
    Result:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        examId:
          type: string
        startTime:
          type: string
          format: date-time
        submitTime:
          type: string
          format: date-time
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerInput'
        correctCount:
          type: integer
        totalQuestions:
          type: integer
        score:
          type: number
        status:
          type: string
          example: completed
    ResultQuestionDetail:
      type: object
      properties:
        questionId:
          type: string
        content:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
        selectedOptionId:
          type: string
          nullable: true
        correctOptionId:
          type: string
        isCorrect:
          type: boolean
        explanation:
          type: string
    ResultDetailed:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            exam:
              $ref: '#/components/schemas/Exam'
            details:
              type: array
              items:
                $ref: '#/components/schemas/ResultQuestionDetail'
...
    AdminDetailedResultQuestion:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
        explanation:
          type: string
        correctOptionId:
          type: string
        selectedOptionId:
          type: string
          nullable: true
        isCorrect:
          type: boolean
    AdminDetailedResult:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            exam:
              $ref: '#/components/schemas/Exam'
            questions:
              type: array
              items:
                $ref: '#/components/schemas/AdminDetailedResultQuestion'